// code-app/src/generated/models/business-group-model.ts
import type { Company } from './company-model';

export interface BusinessGroup {
  /**
   * @displayName Business Group
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Business Group Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  businessGroupName: string;
  /**
   * @displayName Company
   * @validationRule Required for create/update operations
   */
  company: Pick<Company, 'id' | 'companyName'>;
}

export const _BusinessGroup = 'BusinessGroup' as const;